Fix types

+++ core-backend/src/funcs.rs
@@ -94,8 +94,8 @@ pub(crate) fn gas<E: Ext>(ext: LaterExt<E>) -> impl Fn(i32) -> Result<(), &'stat
     }
 }
 
-pub(crate) fn gas_available<E: Ext>(ext: LaterExt<E>) -> impl Fn() -> u64 {
-    move || ext.with(|ext: &mut E| ext.gas_available()).unwrap_or(0)
+pub(crate) fn gas_available<E: Ext>(ext: LaterExt<E>) -> impl Fn() -> i64 {
+    move || ext.with(|ext: &mut E| ext.gas_available()).unwrap_or(0) as i64
 }
 
 pub(crate) fn send_init<E: Ext>(
